refactor(chat-composer): address review — layer boundary, draft key, panel context
- move composerAttachment.ts (pure data) to utils/messageUtils/ so utils/
  composerClipboard imports a sibling, not components/ (fixes the utils->components
  inversion that pulled the dormant composer tree into the live message bundle)
- chatDraftCache: v2 owns 'v2-chat-composer-draft' instead of writing structured
  objects to v1's 'inputbar-draft' string key; migrate the legacy string on read
- panelEngine provider: hand onClose the real memoized context via a ref instead
  of 'context: this' (undefined in an arrow fn)
- panelEngine/index: document the v1 QuickPanel fork + collapse-at-switchover plan
