+++ src/renderer/components/chat/composer/paste/PasteService.ts
@@ -1,5 +1,5 @@
 import { loggerService } from '@logger'
-import { type ComposerAttachment, toComposerAttachment } from '@renderer/components/chat/composer/composerAttachment'
+import { type ComposerAttachment, toComposerAttachment } from '@renderer/utils/messageUtils/composerAttachment'
 import { LONG_TEXT_PASTE_THRESHOLD } from '@renderer/config/constant'
 import { COMPOSER_FILE_KIND, type PastedTextFileMetadata } from '@renderer/types'
 import { getFileExtension, isSupportedFile } from '@renderer/utils'
