+++ src/renderer/components/chat/composer/tools/ComposerToolProvider.tsx
@@ -1,4 +1,4 @@
-import type { ComposerAttachment } from '@renderer/components/chat/composer/composerAttachment'
+import type { ComposerAttachment } from '@renderer/utils/messageUtils/composerAttachment'
 import type { ComposerToolLauncher } from '@renderer/components/chat/composer/toolLauncher'
 import { ensureComposerFileTokenSourceIds } from '@renderer/utils/messageUtils/composerFileTokenSource'
 import type { KnowledgeBase } from '@shared/data/types/knowledge'
