+++ src/renderer/components/chat/composer/tools/definitions/__tests__/attachmentTool.test.tsx
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import type { ComposerAttachment } from '../../../composerAttachment'
+import type { ComposerAttachment } from '@renderer/utils/messageUtils/composerAttachment'
 import type { ComposerSerializedToken } from '../../../tokens'
 import attachmentTool from '../attachmentTool'
