+++ src/renderer/components/chat/composer/variants/__tests__/agentComposerTokens.test.ts
@@ -1,7 +1,7 @@
 import type { LocalSkill } from '@renderer/types'
 import { describe, expect, it } from 'vitest'
 
-import type { ComposerAttachment } from '../../composerAttachment'
+import type { ComposerAttachment } from '@renderer/utils/messageUtils/composerAttachment'
 import {
   agentComposerTokenId,
   agentFileToComposerToken,
