+++ src/renderer/components/chat/composer/variants/__tests__/chatComposerTokens.test.ts
@@ -1,7 +1,7 @@
 import type { KnowledgeBase } from '@shared/data/types/knowledge'
 import { describe, expect, it } from 'vitest'
 
-import type { ComposerAttachment } from '../../composerAttachment'
+import type { ComposerAttachment } from '@renderer/utils/messageUtils/composerAttachment'
 import {
   chatComposerTokenId,
   fileToComposerToken,
