+++ src/renderer/components/chat/composer/variants/shared/__tests__/composerQueuedPayload.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it, vi } from 'vitest'
 
-import type { ComposerAttachment } from '../../../composerAttachment'
+import type { ComposerAttachment } from '@renderer/utils/messageUtils/composerAttachment'
 import type { ComposerSerializedDraft } from '../../../tokens'
 import { buildComposerQueuedPayload } from '../composerQueuedPayload'
