+++ src/renderer/components/chat/composer/variants/shared/composerTokens.ts
@@ -3,7 +3,7 @@ import {
   getComposerFileTokenSourceId
 } from '@renderer/utils/messageUtils/composerFileTokenSource'
 
-import type { ComposerAttachment } from '../../composerAttachment'
+import type { ComposerAttachment } from '@renderer/utils/messageUtils/composerAttachment'
 import type { ComposerDraftToken, ComposerSerializedToken } from '../../tokens'
 
 export const composerFileTokenId = (file: Pick<ComposerAttachment, 'fileTokenSourceId'>) => {
