+++ src/renderer/utils/messageUtils/composerAttachment.ts
